# testing/api/library.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "library.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/books": {
      "get": {
        "operationId": "ListBooks",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiListBooksResponse"
            }
          }
        },
        "tags": [
          "LibraryService"
        ]
      },
      "post": {
        "operationId": "CreateBook",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiBook"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiBook"
            }
          }
        ],
        "tags": [
          "LibraryService"
        ]
      }
    },
    "/books/{book.book_id}": {
      "patch": {
        "operationId": "UpdateBook",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiBook"
            }
          }
        },
        "parameters": [
          {
            "name": "book.book_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiBook"
            }
          }
        ],
        "tags": [
          "LibraryService"
        ]
      }
    },
    "/books/{book_id}": {
      "get": {
        "operationId": "GetBook",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiBook"
            }
          }
        },
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "LibraryService"
        ]
      },
      "patch": {
        "operationId": "DeleteBook",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          }
        },
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "LibraryService"
        ]
      }
    }
  },
  "definitions": {
    "apiBook": {
      "type": "object",
      "properties": {
        "book_id": {
          "type": "string"
        }
      }
    },
    "apiListBooksResponse": {
      "type": "object",
      "properties": {
        "books": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiBook"
          }
        }
      }
    },
    "protobufEmpty": {
      "type": "object",
      "description": "service Foo {\n      rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n    }\n\nThe JSON representation for `Empty` is empty JSON object `{}`.",
      "title": "A generic empty message that you can re-use to avoid defining duplicated\nempty messages in your APIs. A typical example is to use it as the request\nor the response type of an API method. For instance:"
    }
  }
}